move dev to util
